Moved Base64 to a separate ignored util file and documented the return of Response#mime()
