+++ lib/collection/response.js
@@ -1,5 +1,4 @@
 var _ = require('../util').lodash,
-    base64 = require('../util').base64,
     fileType = require('file-type'),
     mime = require('mime-types'),
     Property = require('./property').Property,
@@ -61,7 +60,8 @@ _.extend(Response.prototype, /** @lends Response.prototype */ {
              *
              * @type {Buffer|UInt8Array}
              */
-            rawBody: options.body && !_.isString(options.body) && _.isObject(options.body) ? options.body : undefined,
+            stream: options.body && !_.isString(options.body) && _.isObject(options.body) ?
+                options.body : options.stream,
             /**
              * @type {PropertyList<Cookie>}
              */
@@ -86,15 +86,31 @@ _.extend(Response.prototype, /** @lends Response.prototype */ {
         return this.status || Response.HTTP_REASON_PHRASES[this.code];
     },
 
+    /**
+     * @private
+     *
+     * @param {String} contentType
+     *
+     * @returns {Object}
+     *
+     * @note example object returned
+     * {
+     *   mimeSource: 'header' // or 'content' or 'default'
+     *   mime: string // the content type of response
+     *   ext: string // extension of file that stores this type of data
+     *   base: 'text' // or 'audio', 'video', 'image' ...
+     *   detected: {} // same as root object, but based on what is detected from content
+     * }
+     */
     mime: function (contentType) {
-        var responseBody = this.rawBody == null ? new Uint8Array() : this.rawBody,
+        var responseBody = this.stream == null ? new Uint8Array() : this.stream,
             body;
 
-            body = {
-                mimeSource: 'header',
-                mime: this.headers.one('content-type') && this.headers.one('content-type').value || contentType,
-                detected: fileType(responseBody) || {}
-            };
+        body = {
+            mimeSource: 'header',
+            mime: this.headers.one('content-type') && this.headers.one('content-type').value || contentType,
+            detected: fileType(responseBody) || {}
+        };
 
         // if content type is not resolved, we use the mime to determine the same
         if (!body.mime) {
